Initial commit of ebpf-test.
This utility offloads an eBPF program to a PCI device. If this program
operates on some data, this utility can also copy a binary file to an
NVMe SSD and then repeatedly DMA a (configurable) chunk of this file to
the PCI device and execute the eBPF program on this chunk.